Serialize actions dispatch
This ensures that multiple commands can't simultaneously access the
same input state, in a way that would cause races.

The spec uses a synchronous wait, on the assumption that higher layers
are either serializing commands, or are already running their steps in
parallel, so it's not blocking the "main thread".

This requires any use of actions to go via "dispatch actions" rather
than calling the individual dispatch endpoints, so we need to refactor
some commands that are defined in terms of the actions they create.