Add the Element Origin concept
This is really designed for BiDi, allowing actions to specify the
origin relative to an element in the form:

 {
  type: "element"
  sharedId: <id>
 }

rather than sending in a WebDriver-style element reference. For
simplicit this allows the new construct to work irrespective of
whether actions are being used in a HTTP session or a BiDi
session (but a BiDi session is not expected to support the HTTP-style
web element reference identifiers).