Remove Andreas Tolfsen from the list of code owners.
Andreas didn't contribute nor reviewed PRs for WebDriver
classic for some years. To not get him listed as reviewer
for newly created PRs it would be good to get him removed
from the list of owners.

Andreas, thanks a lot for all the work that you did for
the project!