Require pointerMove and wheel coordinates to be integers
Prevously we allowed these properties to be undefined, and then later
tried to do computations on the values. That doesn't work; they either
need to have an integer default value or require that a value is
supplied.

Since implementations currently seem to raise an exception for these
values being undefined, the easiest approach is to require that they
are supplied, so that's what's implemented here.