Fix setting cursor position in send keys
I believe the expected behaviour is:
* If the target already has focus, do nothing
* Otherwise, if the target is focusable, set the selection to the last
editable point in the element.

Previously the spec was unconditionally focusing the target
element (execpting some cases with file uploads), and therefore the
letter of the spec was to never move the caret.

This change updates it to only focus the target if it's not already
the active element, and then move the caret to the end.