Replace misleading term
This specification relies on the term "response" as defined by the Fetch
standard. The "error" type which is returned by fallible algorithms does
not qualify as such a response. This makes the use of the phrase "error
response" to define the "error" type somewhat confusing.

Replace "error response" with the more generic "error value" to avoid
insinuating a non-existent relationship.