Editorial: remove inaccurate qualifier
This specification uses the List type from WHATWG Infra [1]. It does so
indirectly, working with a derivative which it refers to as a "JSON
List." This type is not defined, but because its usage matches that of
the List, the qualifier "JSON" appears to be superfluous.

Remove the qualifier to clarify the type of the specification values.

[1] https://infra.spec.whatwg.org/#lists