editorial: Use "content navigable" instead of "nested navigable"
The term was renamed in the HTML spec in whatwg/html#8712 so the current
reference was broken as mentioned in #1735.

While here, use ReSpec's shorthand syntax to refer to these sub concepts
directly and avoid having to specifically link to them in the index at the
end of the document.